Rename exceptions.py to exception.py
