+++ bot/api/exception.py
